Import 1.1.93

+++ drivers/scsi/hosts.h
@@ -247,7 +247,7 @@ struct Scsi_Host
 
 		/* These parameters should be set by the detect routine */
 		unsigned char *base;
-		short unsigned int io_port;
+		unsigned int io_port;
 		unsigned char n_io_port;
 		unsigned char irq;
 		unsigned char dma_channel;
@@ -356,3 +356,19 @@ extern void scsi_unregister_module(int, void *);
 #define SG_EXTRA_DEVS (SD_EXTRA_DEVS + SR_EXTRA_DEVS + ST_EXTRA_DEVS)
 
 #endif
+/*
+ * Overrides for Emacs so that we follow Linus's tabbing style.
+ * Emacs will notice this stuff at the end of the file and automatically
+ * adjust the settings for this buffer only.  This must remain at the end
+ * of the file.
+ * ---------------------------------------------------------------------------
+ * Local variables:
+ * c-indent-level: 8
+ * c-brace-imaginary-offset: 0
+ * c-brace-offset: -8
+ * c-argdecl-indent: 8
+ * c-label-offset: -8
+ * c-continued-statement-offset: 8
+ * c-continued-brace-offset: 0
+ * End:
+ */
